use components xml to import inject plugin
