--- src/main/java/io/avaje/inject/mojo/AutoProvidesMojo.java
@@ -1,184 +1,0 @@
-package io.avaje.inject.mojo;
-
-import io.avaje.inject.spi.Module;
-import io.avaje.inject.spi.Plugin;
-import org.apache.maven.artifact.Artifact;
-import org.apache.maven.artifact.resolver.filter.ScopeArtifactFilter;
-import org.apache.maven.plugin.AbstractMojo;
-import org.apache.maven.plugin.MojoExecutionException;
-import org.apache.maven.plugin.logging.Log;
-import org.apache.maven.plugins.annotations.LifecyclePhase;
-import org.apache.maven.plugins.annotations.Mojo;
-import org.apache.maven.plugins.annotations.Parameter;
-import org.apache.maven.plugins.annotations.ResolutionScope;
-import org.apache.maven.project.MavenProject;
-
-import static java.util.stream.Collectors.joining;
-import static java.util.stream.Collectors.toList;
-
-import java.io.File;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.lang.reflect.Type;
-import java.net.MalformedURLException;
-import java.net.URL;
-import java.net.URLClassLoader;
-import java.util.*;
-
-/**
- * Plugin that generates <code>target/avaje-module-provides.txt</code> and <code>
- * target/avaje-plugin-provides.txt</code> based on the avaje-inject modules and plugins in the
- * classpath.
- *
- * <p>This allows the avaje-inject-generator annotation processor to be aware of all the components
- * and plugins provided by other modules in the classpath at compile time.
- */
-@Mojo(
-    name = "provides",
-    defaultPhase = LifecyclePhase.PROCESS_SOURCES,
-    requiresDependencyResolution = ResolutionScope.COMPILE)
-public class AutoProvidesMojo extends AbstractMojo {
-
-  private static final String DISABLING_AVAJE_MODULE_VERIFICATION =
-      "disabling avaje module verification";
-
-  @Parameter(defaultValue = "${project}", readonly = true, required = true)
-  private MavenProject project;
-
-  private final List<AvajeModule> modules = new ArrayList<>();
-
-  @Override
-  public void execute() throws MojoExecutionException {
-    final var listUrl = compileDependencies();
-
-    final var directory = new File(project.getBuild().getDirectory());
-    if (!directory.exists()) {
-      directory.mkdirs();
-    }
-
-    try (var newClassLoader = createClassLoader(listUrl);
-        var flagFile = createFileWriter("avaje-plugin-exists.txt");
-        var moduleWriter = createFileWriter("avaje-module-provides.txt");
-        var pluginWriter = createFileWriter("avaje-plugin-provides.txt");
-        var moduleCSV = createFileWriter("avaje-module-dependencies.csv")) {
-
-      flagFile.append(DISABLING_AVAJE_MODULE_VERIFICATION);
-      getLog().info(DISABLING_AVAJE_MODULE_VERIFICATION);
-      writeProvidedPlugins(newClassLoader, pluginWriter);
-      writeProvidedModules(newClassLoader, moduleWriter);
-      writeModuleCSV(moduleCSV);
-
-    } catch (final IOException e) {
-      throw new MojoExecutionException("Failed to write spi classes", e);
-    }
-  }
-
-  private List<URL> compileDependencies() throws MojoExecutionException {
-    final List<URL> listUrl = new ArrayList<>();
-    project.setArtifactFilter(new ScopeArtifactFilter("compile"));
-    final var deps = project.getArtifacts();
-
-    for (final Artifact artifact : deps) {
-      try {
-        listUrl.add(artifact.getFile().toURI().toURL());
-      } catch (final MalformedURLException e) {
-        throw new MojoExecutionException("Failed to get compile dependencies", e);
-      }
-    }
-    return listUrl;
-  }
-
-  private URLClassLoader createClassLoader(List<URL> listUrl) {
-    return new URLClassLoader(
-        listUrl.toArray(new URL[listUrl.size()]), Thread.currentThread().getContextClassLoader());
-  }
-
-  private FileWriter createFileWriter(String string) throws IOException {
-    return new FileWriter(new File(project.getBuild().getDirectory(), string));
-  }
-
-  private void writeProvidedPlugins(URLClassLoader newClassLoader, FileWriter pluginWriter)
-      throws IOException {
-    final Set<String> providedTypes = new HashSet<>();
-
-    final Log log = getLog();
-    for (final var plugin : ServiceLoader.load(Plugin.class, newClassLoader)) {
-      log.info("Loaded Plugin: " + plugin.getClass().getTypeName());
-
-      for (final Type provide : plugin.provides()) {
-        providedTypes.add(provide.getTypeName());
-      }
-      for (final Type provide : plugin.providesAspects()) {
-        providedTypes.add(wrapAspect(provide.getTypeName()));
-      }
-    }
-
-    for (final var providedType : providedTypes) {
-      pluginWriter.write(providedType);
-      pluginWriter.write("\n");
-    }
-  }
-
-  private void writeProvidedModules(URLClassLoader newClassLoader, FileWriter moduleWriter)
-      throws IOException {
-    final Set<String> providedTypes = new HashSet<>();
-
-    final Log log = getLog();
-    for (final var module : ServiceLoader.load(Module.class, newClassLoader)) {
-
-      final var name = module.getClass().getTypeName();
-      log.info("Detected External Module: " + name);
-
-      final var provides = new ArrayList<String>();
-      for (final Type provide : module.provides()) {
-        var type = provide.getTypeName();
-        providedTypes.add(type);
-        provides.add(type);
-      }
-      for (final Type provide : module.autoProvides()) {
-        var type = provide.getTypeName();
-        providedTypes.add(type);
-        provides.add(type);
-      }
-      for (final Type provide : module.autoProvidesAspects()) {
-        var type = wrapAspect(provide.getTypeName());
-        providedTypes.add(type);
-        provides.add(type);
-      }
-
-      final var requires =
-          Arrays.<Type>stream(module.requires()).map(Type::getTypeName).collect(toList());
-
-      Arrays.<Type>stream(module.autoRequires()).map(Type::getTypeName).forEach(requires::add);
-      Arrays.<Type>stream(module.requiresPackages()).map(Type::getTypeName).forEach(requires::add);
-      Arrays.<Type>stream(module.autoRequiresAspects())
-          .map(Type::getTypeName)
-          .map(AutoProvidesMojo::wrapAspect)
-          .forEach(requires::add);
-      modules.add(new AvajeModule(name, provides, requires));
-    }
-
-    for (final String providedType : providedTypes) {
-      moduleWriter.write(providedType);
-      moduleWriter.write("\n");
-    }
-  }
-
-  private void writeModuleCSV(FileWriter moduleWriter) throws IOException {
-    moduleWriter.write("External Module Type|Provides|Requires");
-    for (AvajeModule avajeModule : modules) {
-      moduleWriter.write("\n");
-      moduleWriter.write(avajeModule.name());
-      moduleWriter.write("|");
-      var provides = avajeModule.provides().stream().collect(joining(","));
-      moduleWriter.write(provides.isEmpty() ? " " : provides);
-      moduleWriter.write("|");
-      var requires = avajeModule.requires().stream().collect(joining(","));
-      moduleWriter.write(requires.isEmpty() ? " " : requires);
-    }
-  }
-
-  private static String wrapAspect(String aspect) {
-    return "io.avaje.inject.aop.AspectProvider<" + aspect + ">";
-  }
-}
