--- src/main/java/io/avaje/inject/mojo/AvajeModule.java
@@ -1,5 +1,0 @@
-package io.avaje.inject.mojo;
-
-import java.util.List;
-
-record AvajeModule(String name, List<String> provides, List<String> requires) {}
